style: post 요청 분리
